+++ notification-service/src/main/java/com/muazwzxv/notifications/NotificationRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 import java.util.UUID;
 
-public interface NotificationRepository extends JpaRepository<Notifications, Integer> {
+public interface NotificationRepository extends JpaRepository<Notifications, Long> {
     Optional<Notifications> findByUuid(UUID uuid);
 }
